Feat: Criando uma classe de erros personalizada

+++ gameteams/src/storage/group/groupCreate.ts
@@ -2,13 +2,19 @@ import AsyncStorage from '@react-native-async-storage/async-storage';
 
 import { GROUP_COLLECTION } from '@storage/storageConfig';
 import { groupsGetAll } from './groupGetAll';
+import { AppError } from '@utils/AppError';
 
 export async function groupCreate(newGroup: string) {
   try {
     const storedGroups = await groupsGetAll();
+
+    const groupAlreadyExists = storedGroups.includes(newGroup)
+
+    if(groupAlreadyExists){
+      throw new AppError('Já exsite um grupo cadastrado com esse nome')
+    }
+
     const storage = JSON.stringify([...storedGroups, newGroup])
-    
-                              // Chave, valor => Precisa ser uma string
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
   } catch (error) {
     throw error;
